Moved "isSundayTheObserverSubscription" method to a Shared directory
